[refactor] UI: combine adaptive composables and use sample data for previews
