+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/account/OnboardingScreen.kt
@@ -29,13 +29,13 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import com.rwmobi.kunigami.domain.model.Tariff
 import com.rwmobi.kunigami.domain.model.account.Account
 import com.rwmobi.kunigami.domain.model.account.Agreement
 import com.rwmobi.kunigami.domain.model.account.ElectricityMeterPoint
 import com.rwmobi.kunigami.domain.model.account.UserProfile
 import com.rwmobi.kunigami.ui.destinations.account.components.AppInfoFooter
 import com.rwmobi.kunigami.ui.destinations.account.components.CredentialsInputForm
+import com.rwmobi.kunigami.ui.previewsampledata.TariffSamples
 import com.rwmobi.kunigami.ui.theme.AppTheme
 import com.rwmobi.kunigami.ui.theme.getDimension
 import kotlinx.datetime.Clock
@@ -158,17 +158,7 @@ private fun Preview() {
                                 ),
                             ),
                         ),
-                        tariff = Tariff(
-                            productCode = "AGILE-FLEX-22-11-25",
-                            tariffCode = "E-1R-AGILE-FLEX-22-11-25-A",
-                            fullName = "Octopus 12M Fixed April 2024 v1",
-                            displayName = "Octopus 12M Fixed",
-                            vatInclusiveUnitRate = 99.257,
-                            vatInclusiveStandingCharge = 94.682,
-                            description = "With Agile Octopus, you get access to half-hourly energy prices, tied to wholesale prices and updated daily.  The unit rate is capped at 100p/kWh (including VAT).",
-                            availableFrom = Clock.System.now(),
-                            availableTo = null,
-                        ),
+                        tariff = TariffSamples.agileFlex221125,
                     ),
                     errorMessages = listOf(),
                 ),
