+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/account/components/ClearCredentialSectionAdaptive.kt
@@ -9,6 +9,7 @@ package com.rwmobi.kunigami.ui.destinations.account.components
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -16,6 +17,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -23,6 +25,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
 import com.rwmobi.kunigami.ui.components.IconTextButton
 import com.rwmobi.kunigami.ui.theme.AppTheme
 import com.rwmobi.kunigami.ui.theme.getDimension
@@ -35,7 +38,65 @@ import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-internal fun ClearCredentialSectionWide(
+internal fun ClearCredentialSectionAdaptive(
+    modifier: Modifier = Modifier,
+    onClearCredentialButtonClicked: () -> Unit,
+    useWideLayout: Boolean = false,
+) {
+    if (useWideLayout) {
+        ClearCredentialSectionWide(
+            modifier = modifier,
+            onClearCredentialButtonClicked = onClearCredentialButtonClicked,
+        )
+    } else {
+        ClearCredentialSectionCompact(
+            modifier = modifier,
+            onClearCredentialButtonClicked = onClearCredentialButtonClicked,
+        )
+    }
+}
+
+@Composable
+private fun ClearCredentialSectionCompact(
+    modifier: Modifier = Modifier,
+    onClearCredentialButtonClicked: () -> Unit,
+) {
+    val dimension = LocalDensity.current.getDimension()
+
+    Column(
+        modifier = modifier
+            .clip(shape = MaterialTheme.shapes.medium)
+            .background(color = MaterialTheme.colorScheme.surfaceContainer)
+            .padding(all = dimension.grid_2),
+    ) {
+        Text(
+            style = MaterialTheme.typography.titleMedium,
+            color = MaterialTheme.colorScheme.onSurface,
+            fontWeight = FontWeight.Bold,
+            text = stringResource(resource = Res.string.account_clear_credential_title),
+        )
+
+        Spacer(modifier = Modifier.size(size = dimension.grid_2))
+
+        Text(
+            color = MaterialTheme.colorScheme.onSurface,
+            style = MaterialTheme.typography.bodyMedium,
+            text = stringResource(resource = Res.string.account_clear_credential_description),
+        )
+
+        Spacer(modifier = Modifier.size(size = dimension.grid_4))
+
+        IconTextButton(
+            modifier = Modifier.align(alignment = Alignment.End),
+            icon = painterResource(resource = Res.drawable.eraser),
+            text = stringResource(resource = Res.string.account_clear_credential_button_cta),
+            onClick = onClearCredentialButtonClicked,
+        )
+    }
+}
+
+@Composable
+private fun ClearCredentialSectionWide(
     modifier: Modifier = Modifier,
     onClearCredentialButtonClicked: () -> Unit,
 ) {
@@ -79,11 +140,24 @@ internal fun ClearCredentialSectionWide(
 
 @Preview
 @Composable
-private fun ClearCredentialSectionPreview() {
+private fun Preview() {
     AppTheme {
-        ClearCredentialSectionWide(
-            modifier = Modifier.fillMaxWidth(),
-            onClearCredentialButtonClicked = {},
-        )
+        Surface(modifier = Modifier.padding(all = 16.dp)) {
+            Column(
+                verticalArrangement = Arrangement.spacedBy(space = 16.dp),
+            ) {
+                ClearCredentialSectionAdaptive(
+                    modifier = Modifier.fillMaxWidth(),
+                    onClearCredentialButtonClicked = {},
+                    useWideLayout = false,
+                )
+
+                ClearCredentialSectionAdaptive(
+                    modifier = Modifier.fillMaxWidth(),
+                    onClearCredentialButtonClicked = {},
+                    useWideLayout = true,
+                )
+            }
+        }
     }
 }
